requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
python-nmap==0.7.1
scapy==2.4.5
pytest==7.4.0
